requirements.txt
discord.py==1.7.3
feedparser==6.0.6